+++ .flake8
@@ -12,3 +12,4 @@ exclude =
 per-file-ignores =
     ../Fink_GRB/fink_grb/online/ztf_join_gcn.py:W503,E402
     ../Fink_GRB/fink_grb/offline/spark_offline.py:W503,W605
+    ../Fink_GRB/fink_grb/utils/fun_utils.py:F811
