+++ fink_grb/utils/grb_prob.py
@@ -1,6 +1,14 @@
 import numpy as np
+import pandas as pd
 from math import pi
 from scipy.stats import poisson
+
+from pyspark.sql.functions import pandas_udf, col  # noqa: F401
+from pyspark.sql.types import DoubleType
+
+from astropy.coordinates import SkyCoord
+import astropy.units as u
+from astropy.time import Time
 
 
 def p_ser_grb_vect(
@@ -101,16 +109,146 @@ def p_ser_grb_vect(
     return p_sers
 
 
+@pandas_udf(DoubleType())
+def grb_assoc(
+    ztf_ra: pd.Series,
+    ztf_dec: pd.Series,
+    jdstarthist: pd.Series,
+    platform: pd.Series,
+    trigger_time: pd.Series,
+    grb_ra: pd.Series,
+    grb_dec: pd.Series,
+    grb_error: pd.Series,
+) -> pd.Series:
+    """
+    Find the ztf alerts falling in the error box of the notices and emits after the trigger time.
+    Then, Compute an association serendipitous probability for each of them and return it.
+
+    Parameters
+    ----------
+    ztf_ra : double spark column
+        right ascension coordinates of the ztf alerts
+    ztf_dec : double spark column
+        declination coordinates of the ztf alerts
+    jdstarthist : double spark column
+        Earliest Julian date of epoch corresponding to ndethist [days]
+        ndethist : Number of spatially-coincident detections falling within 1.5 arcsec
+            going back to beginning of survey;
+            only detections that fell on the same field and readout-channel ID
+            where the input candidate was observed are counted.
+            All raw detections down to a photometric S/N of ~ 3 are included.
+    platform : string spark column
+        voevent emitting platform
+    trigger_time : double spark column
+        grb trigger time (UTC)
+    grb_ra : double spark column
+        grb right ascension
+    grb_dec : double spark column
+        grb declination
+    grb_error : double spark column
+        grb error radius (in arcminute)
+
+    Returns
+    grb_proba : pandas Series
+        the serendipitous probability for each ztf alerts.
+
+    Examples
+    --------
+
+    >>> sparkDF = spark.read.format('parquet').load(join_data)
+
+    >>> df_grb = sparkDF.withColumn(
+    ... "grb_proba",
+    ... grb_assoc(
+    ...    sparkDF.candidate.ra,
+    ...     sparkDF.candidate.dec,
+    ...     sparkDF.candidate.jdstarthist,
+    ...     sparkDF.platform,
+    ...     sparkDF.timeUTC,
+    ...     sparkDF.ra,
+    ...     sparkDF.dec,
+    ...     sparkDF.err
+    ...  ),
+    ... )
+
+    >>> df_grb = df_grb.select([
+    ... "objectId",
+    ... "candid",
+    ... col("candidate.ra").alias("ztf_ra"),
+    ... col("candidate.dec").alias("ztf_dec"),
+    ... "candidate.jd",
+    ... "platform",
+    ... "instrument",
+    ... "trigger_id",
+    ... col("ra").alias("grb_ra"),
+    ... col("dec").alias("grb_dec"),
+    ... col("err_arcmin").alias("grb_loc_error"),
+    ... "timeUTC",
+    ... "grb_proba"
+    ... ])
+
+    >>> grb_prob = df_grb.toPandas()
+    >>> grb_test = pd.read_parquet("fink_grb/test/test_data/grb_prob_test.parquet")
+    >>> assert_frame_equal(grb_prob, grb_test)
+    """
+    grb_proba = np.ones_like(ztf_ra.values, dtype=float) * -1.0
+    platform = platform.values
+
+    # array of events detection rates in events/years
+    # depending of the instruments
+    condition = [
+        np.equal(platform, "Fermi"),
+        np.equal(platform, "SWIFT"),
+        np.equal(platform, "INTEGRAL"),
+        np.equal(platform, "ICECUBE"),
+    ]
+    choice_grb_rate = [250, 100, 60, 8]
+    grb_det_rate = np.select(condition, choice_grb_rate)
+
+    # array of error box
+    grb_error = grb_error.values
+
+    trigger_time = Time(
+        pd.to_datetime(trigger_time.values, utc=True), format="datetime"
+    ).jd
+
+    # alerts emits after the grb
+    delay = jdstarthist - trigger_time
+    time_condition = delay > 0
+
+    ztf_coords = SkyCoord(ztf_ra, ztf_dec, unit=u.degree)
+    grb_coord = SkyCoord(grb_ra, grb_dec, unit=u.degree)
+
+    # alerts falling within the grb_error_box
+    spatial_condition = (
+        ztf_coords.separation(grb_coord).arcminute < 1.5 * grb_error
+    )  # 63.5 * grb_error
+
+    # convert the delay in year
+    delay_year = delay[time_condition & spatial_condition] / 365.25
+
+    # compute serendipitous probability
+    p_ser = p_ser_grb_vect(
+        grb_error[time_condition & spatial_condition] / 60,
+        delay_year.values,
+        grb_det_rate[time_condition & spatial_condition],
+    )
+
+    grb_proba[time_condition & spatial_condition] = p_ser[0]
+
+    return pd.Series(grb_proba)
+
+
 if __name__ == "__main__":  # pragma: no cover
-    import sys
-    import doctest
     import pandas as pd  # noqa: F401
-    from astropy.time import Time  # noqa: F401
     from math import sqrt  # noqa: F401
     from scipy import special  # noqa: F401
+    from fink_utils.test.tester import spark_unit_tests_science
+    from pandas.testing import assert_frame_equal  # noqa: F401
 
-    if "unittest.util" in __import__("sys").modules:
-        # Show full diff in self.assertEqual.
-        __import__("sys").modules["unittest.util"]._MAX_LENGTH = 999999999
+    globs = globals()
 
-    sys.exit(doctest.testmod()[0])
+    join_data = "fink_grb/test/test_data/join_raw_datatest.parquet"
+
+    # Run the test suite
+    spark_unit_tests_science(globs)
